# docs/openapi/teamclaw-api.v1.yaml
openapi: 3.1.0
info:
  title: TeamClaw Cloud API
  version: 1.0.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: https://api.teamclaw.local
security:
  - bearerAuth: []
...
paths:
...
  /v1/sessions/{sessionId}/messages:
    get:
      operationId: listMessages
      summary: List messages
      tags: [Messages]
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Session messages.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                type: object
                required: [items, nextCursor]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Message"
                  nextCursor:
                    type:
                      - string
                      - "null"
        "401":
          $ref: "#/components/responses/Error"
    post:
      operationId: insertMessage
      summary: Insert message
      tags: [Messages]
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - name: Idempotency-Key
          in: header
          schema:
            type: string
            minLength: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InsertMessageRequest"
      responses:
        "200":
          description: Inserted or replayed message.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /v1/messages/{messageId}:
    patch:
      operationId: patchMessage
      summary: Update message content or metadata
      tags: [Messages]
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                metadata:
                  type:
                    - object
                    - "null"
                  additionalProperties: true
      responses:
        "200":
          description: Updated message.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
    delete:
      operationId: deleteMessage
      summary: Delete message
      tags: [Messages]
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: OK
        "401":
          $ref: "#/components/responses/Error"
...
  /v1/teams/{teamId}/shortcuts:
    get:
      operationId: listShortcuts
      summary: List team shortcuts
      tags: [Shortcuts]
      parameters:
        - $ref: "#/components/parameters/TeamId"
        - name: parentId
          in: query
          schema:
            type:
              - string
              - "null"
            format: uuid
      responses:
        "200":
          description: Team shortcuts.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Shortcut"
        "401":
          $ref: "#/components/responses/Error"
  /v1/shortcuts:
    post:
      operationId: createShortcut
      summary: Create shortcut
      tags: [Shortcuts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShortcutCreate"
      responses:
        "201":
          description: Created shortcut.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shortcut"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /v1/shortcuts/{shortcutId}:
    patch:
      operationId: updateShortcut
      summary: Update shortcut
      tags: [Shortcuts]
      parameters:
        - name: shortcutId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShortcutPatch"
      responses:
        "200":
          description: Updated shortcut.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shortcut"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
    delete:
      operationId: deleteShortcut
      summary: Delete shortcut
      tags: [Shortcuts]
      parameters:
        - name: shortcutId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Shortcut deleted.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
        "401":
          $ref: "#/components/responses/Error"
  /v1/shortcuts/batch-move:
    post:
      operationId: batchMoveShortcuts
      summary: Batch move shortcuts
      tags: [Shortcuts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [moves]
              properties:
                moves:
                  type: array
                  items:
                    type: object
                    required: [shortcutId, parentId, position]
                    properties:
                      shortcutId:
                        type: string
                        format: uuid
                      parentId:
                        type:
                          - string
                          - "null"
                        format: uuid
                      position:
                        type: integer
      responses:
        "204":
          description: Shortcuts moved.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /v1/shortcuts/{shortcutId}/visible-roles:
    put:
      operationId: setShortcutVisibleRoles
      summary: Set shortcut visible roles
      tags: [Shortcuts]
      parameters:
        - name: shortcutId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [roleIds]
              properties:
                roleIds:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        "204":
          description: Visible roles set.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /v1/teams/{teamId}/roles:
    get:
      operationId: listTeamRoles
      summary: List team roles
      tags: [Teams]
      parameters:
        - $ref: "#/components/parameters/TeamId"
      responses:
        "200":
          description: Team roles.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/TeamRole"
        "401":
          $ref: "#/components/responses/Error"
  /v1/teams/{teamId}/permissions:
    get:
      operationId: listTeamPermissions
      summary: List team permissions
      tags: [Teams]
      parameters:
        - $ref: "#/components/parameters/TeamId"
      responses:
        "200":
          description: Team permissions.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Permission"
        "401":
          $ref: "#/components/responses/Error"
...
  /v1/feedback:
    post:
      operationId: submitFeedback
      summary: Submit feedback for a message
      tags: [Telemetry]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [messageId, actorId, kind]
              properties:
                messageId:
                  type: string
                  format: uuid
                actorId:
                  type: string
                  format: uuid
                kind:
                  type: string
                  enum: [up, down, star]
                starRating:
                  type:
                    - integer
                    - "null"
                  minimum: 1
                  maximum: 5
                note:
                  type:
                    - string
                    - "null"
      responses:
        "201":
          description: Feedback submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feedback"
        "401":
          $ref: "#/components/responses/Error"
    get:
      operationId: listFeedback
      summary: List feedback for a session
      tags: [Telemetry]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: sessionId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Feedback list
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Feedback"
        "401":
          $ref: "#/components/responses/Error"
  /v1/feedback/{messageId}:
    delete:
      operationId: deleteFeedback
      summary: Delete feedback for a message
      tags: [Telemetry]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: messageId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: OK
        "401":
          $ref: "#/components/responses/Error"
  /v1/teams/{teamId}/leaderboard:
    get:
      operationId: getTeamLeaderboard
      summary: Get the leaderboard for a team
      tags: [Telemetry]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: teamId
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: period
          schema:
            type: string
            enum: [day, week, month]
            default: week
      responses:
        "200":
          description: Leaderboard
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/LeaderboardEntry"
        "401":
          $ref: "#/components/responses/Error"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  headers:
    RequestId:
      schema:
        type: string
  parameters:
    TeamId:
      name: teamId
      in: path
      required: true
      schema:
        type: string
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
  responses:
...
    Error:
      description: Error response.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code:
              type: string
              enum:
                - missing_auth
                - invalid_json
                - validation_failed
                - forbidden
                - not_found
                - conflict
                - rate_limited
                - upstream_unavailable
                - internal
            message:
              type: string
            requestId:
              type: string
...
    Message:
      type: object
      required: [id, teamId, sessionId, kind, content, createdAt]
      properties:
        id:
          type: string
        teamId:
          type: string
        sessionId:
          type: string
        turnId:
          type:
            - string
            - "null"
        senderActorId:
          type:
            - string
            - "null"
        replyToMessageId:
          type:
            - string
            - "null"
        kind:
          type: string
        content:
          type: string
        metadata:
          type:
            - object
            - "null"
        model:
          type:
            - string
            - "null"
        createdAt:
          type: string
        updatedAt:
          type:
            - string
            - "null"
    InsertMessageRequest:
      type: object
      required: [id, teamId, senderActorId, content]
      properties:
        id:
          type: string
        teamId:
          type: string
        senderActorId:
          type: string
        content:
          type: string
        kind:
          type: string
          default: text
        metadata:
          type:
            - object
            - "null"
        turnId:
          type:
            - string
            - "null"
        replyToMessageId:
          type:
            - string
            - "null"
        model:
          type:
            - string
            - "null"
        createdAt:
          type:
            - string
            - "null"
...
    Shortcut:
      type: object
      required: [id, teamId, kind, label, position]
      properties:
        id:
          type: string
          format: uuid
        teamId:
          type: string
          format: uuid
        parentId:
          type:
            - string
            - "null"
          format: uuid
        kind:
          type: string
        label:
          type: string
        payload:
          type:
            - object
            - "null"
          additionalProperties: true
        position:
          type: integer
        visibleRoleIds:
          type: array
          items:
            type: string
            format: uuid
        createdAt:
          type:
            - string
            - "null"
          format: date-time
        updatedAt:
          type:
            - string
            - "null"
          format: date-time
    ShortcutCreate:
      type: object
      required: [teamId, kind, label]
      properties:
        id:
          type: string
          format: uuid
        teamId:
          type: string
          format: uuid
        parentId:
          type:
            - string
            - "null"
          format: uuid
        kind:
          type: string
        label:
          type: string
        payload:
          type:
            - object
            - "null"
          additionalProperties: true
        position:
          type: integer
        visibleRoleIds:
          type: array
          items:
            type: string
            format: uuid
    ShortcutPatch:
      type: object
      properties:
        parentId:
          type:
            - string
            - "null"
          format: uuid
        label:
          type: string
        payload:
          type:
            - object
            - "null"
          additionalProperties: true
        position:
          type: integer
        visibleRoleIds:
          type: array
          items:
            type: string
            format: uuid
    TeamRole:
      type: object
      required: [id, teamId, code, name]
      properties:
        id:
          type: string
          format: uuid
        teamId:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
    Permission:
      type: object
      required: [resourceId, roleIds]
      properties:
        resourceId:
          type: string
        roleIds:
          type: array
          items:
            type: string
            format: uuid
...
    Feedback:
      type: object
      required: [messageId, actorId, kind, createdAt]
      properties:
        messageId:
          type: string
          format: uuid
        actorId:
          type: string
          format: uuid
        kind:
          type: string
          enum: [up, down, star]
        starRating:
          type:
            - integer
            - "null"
          minimum: 1
          maximum: 5
        note:
          type:
            - string
            - "null"
        createdAt:
          type:
            - string
            - "null"
          format: date-time
        updatedAt:
          type:
            - string
            - "null"
          format: date-time
    LeaderboardEntry:
      type: object
      required: [actorId, period, score]
      properties:
        actorId:
          type: string
          format: uuid
        teamId:
          type:
            - string
            - "null"
          format: uuid
        period:
          type: string
          enum: [day, week, month]
        score:
          type: number
        rank:
          type:
            - integer
            - "null"
        displayName:
          type:
            - string
            - "null"